chore: hoist request types up
